Re-organize agent code
