+++ typescript/packages/lookslike-prototype/src/agent/plan.ts
@@ -1,174 +1,12 @@
 import {
   ChatCompletionMessage,
-  ChatCompletionMessageParam,
-  ChatCompletionTool
+  ChatCompletionMessageParam
 } from "openai/resources/index.mjs";
-import { client, grabJson, messageReducer, model, toolSpec } from "./llm.js";
+import { client, grabJson, messageReducer, model } from "./llm.js";
 import { recordThought, suggestions, updateThought } from "./model.js";
-import { Recipe } from "./data.js";
-
-export const codePrompt = `
-  Your task is to take a user description or request and produce a series of nodes for a computation graph. Nodes can be code blocks or UI components and they communicate with named ports.
-
-  You will construct the graph using the available tools to add, remove, replace and list nodes.
-  You will provide the required edges to connect data from the environment to the inputs of the node. The keys of \`in\` are the names of local inputs and the values are NodePaths (of the form [context, nodeId], where context is typically '.' meaning local namespace).
-
-  "Imagine some todos" ->
-
-  addCodeNode({
-    "id": "todos",
-    "code": "return [{ label: 'Water my plants', checked: false }, { label: 'Buy milk', checked: true }];"
-  })
-
-  Tasks that take no inputs require no edges.
-  All function bodies must take zero parameters. Inputs can be accessed via 'read' and 'deref'.
-
-  ---
-
-  "Remind me to water the plants" ->
-
-  addCodeNode({
-    "id": "addReminder",
-    "code": "const todos = input('todos');\nconst newTodo = { label: 'water the plants', checked: false };\nconst newTodos = [...todos, newTodo];\nreturn newTodos;"
-  })
-
-  Tasks that take no inputs require no edges.
-
-  ---
-
-
-  "Take the existing todos and filter to unchecked" ->
-
-  addCodeNode({
-    "id": "filteredTodos",
-    "code": "const todos = input('todos');\nreturn todos.filter(todo => todo.checked);"
-  })
-
-  Tasks that filter other data must pipe the data through the edges.
-  All function bodies must take zero parameters. Inputs can be accessed via 'input()', values may be null.
-  Always respond with code, even for static data. Wrap your response in a json block. Respond with nothing else.
-
-  ---
-
-  "render each image by url" ->
-  The output of a code node will be bound to the input named 'images'
-
-  addUiNode({
-    "id": "imageUi",
-    "uiTree": {
-      "tag": "ul",
-      "props": {
-        "className": "image"
-      },
-      "children": [
-        "type": "repeat",
-        "binding": "images",
-        "template": {
-          "tag": "li",
-          "props": {},
-          "children": [
-            {
-              "tag": "img",
-              "props": {
-                "src": { "@type": 'binding', "name": 'src' },
-              },
-              "children": []
-            }
-          ],
-        }
-      ]
-    }
-  })
-
-  ---
-
-  "show some text" ->
-  The output of a code node will be bound to the input named 'text'
-
-  addUiNode({
-    "id": "imageUi",
-    "uiTree": {
-      "tag": "span",
-      "props": {
-        "innerText": { "@type": "binding", "name": "text" }
-      },
-      "children": []
-    }
-  })
-
-  ---
-
-  "make a clickable button" ->
-
-  addEventNode({
-    "id": "clickEvent"
-  })
-
-  addUiNode({
-    "id": "buttonUi",
-    "uiTree": {
-      "tag": "button",
-      "props": {
-        "innerText": "Click me",
-        "@click": { "@type": "binding", "name": "onClicked"}
-      },
-      "children": []
-    }
-  })
-
-  addConnection({
-    "from": "clickEvent",
-    "to": ["buttonUi", "onClicked"]
-  })
-
-  ---
-
-  "render my todos" ->
-  The output of a code node will be bound to the input named 'todos'
-
-  addUiNode({
-    "id": "todoUi",
-    "uiTree": {
-      "tag": "ul",
-      "props": {
-        "className": "todo"
-      },
-      "children": [
-        {
-          "@type": "repeat",
-          "name": "todos",
-          "template": {
-            "tag": "li",
-            "props": {},
-            "children": [
-              {
-                "tag": "input",
-                "props": {
-                  "type": "checkbox",
-                  "checked": { "@type": "binding", "name": "checked" }
-                },
-                "children": []
-              },
-              {
-                "tag": "span",
-                "props": {
-                  "className": "todo-label",
-                  "innerText": { "@type": "binding", "name": "label" }
-                },
-                "children": []
-              }
-            ]
-          }
-        }
-      ]
-    }
-  })
-
-  UI trees cannot use any javascript methods, code blocks must prepare the data for the UI to consume.
-  Plain text nodes MUST be applied using the innerText prop in a span, you cannot use a raw string.
-  GLSL shaders cannot declare uniforms other than iTime, iResolution, iMouse and iChannel0 (the user's webcam).
-  notalk;justgo
-`;
+import { Recipe } from "../data.js";
+import { codePrompt } from "./implement.js";
+import { describeTools, toolSpec } from "./tools.js";
 
 type Conversation = ChatCompletionMessageParam[];
 
@@ -256,28 +94,6 @@ export async function suggest(input: string, fullPlan: Conversation) {
   }
 
   return response.choices[0].message;
-}
-
-export function describeTools(
-  tools: ChatCompletionTool[],
-  includeParameters: boolean = false
-) {
-  return tools
-    .map((tool) => {
-      const description = `- ${tool.function.name}: ${tool.function.description}`;
-      const properties = Object.entries(
-        tool.function.parameters?.properties || {}
-      )
-        .map(([name, { type, description }]) => {
-          return `  - ${name} (${type}): ${description}`;
-        })
-        .join("\n");
-      if (!includeParameters) {
-        return description;
-      }
-      return `${description}\n${properties}`;
-    })
-    .join("\n");
 }
 
 export function prepareSteps(userInput: string, recipe: Recipe) {
