+++ typescript/packages/lookslike-prototype/src/graph.ts
@@ -4,7 +4,6 @@ import { Recipe, RecipeNode } from "./data.js";
 
 import { signal, config } from "@commontools/common-frp";
 import { Context } from "./state.js";
-import { generateImage, streamLlm } from "./llm.js";
 import {
   CONTENT_TYPE_CLOCK,
   CONTENT_TYPE_EVENT,
